requirements.txt
Django==4.0.3
tensorflow
keras
numpy
Pillow
matplotlib
IPython
scikit-learn
joblib
